logout bugfix, links, trophy pictures

+++ Web/project/server/routes/trophies.js
@@ -3,6 +3,9 @@ const router = require(`express`).Router()
 const firestore = require(`../config/db`);
 const middleware = require(`../middleware/middleware`);
 
+const path = require("path");
+const multer = require("multer");
+
 const trophyRef = firestore.db.collection('trophy');
  
 
@@ -28,7 +31,6 @@ router.post(`/trophiesList/`, middleware.isLogged, middleware.isAdmin, async (re
                 name: doc.data().name,
                 description: doc.data().description,
                 cost: doc.data().cost,
-                image: doc.data().image,
                 _id: doc.id
             }
             trophyTable.push(temp);
@@ -71,8 +73,29 @@ router.get(`/trophy/:id`,  middleware.isLogged, middleware.isAdmin, async (req, 
             name: doc.data().name,
             description: doc.data().description,
             cost: doc.data().cost,
-            image: doc.data().image,
+			pic: "defaultTrophy.jpg"
+			
+            , _id: doc.id
+			
         }
+		
+		// img
+		const pic = doc.id;
+		
+		console.log(`Pic: ${id}`);
+			
+		let bucket = firestore.admin.storage().bucket();
+		//console.log(bucket);
+		try {
+			await bucket.file("trophies/" + pic).download({destination: "./public/trophies/" + pic}); 
+			
+			trophyObject.pic = pic;
+		} catch (error) {
+			console.log("no pic");
+			console.log(error);
+		}
+		
+		console.log(trophyObject);
 
 
 		res.json(trophyObject);
@@ -105,6 +128,68 @@ router.put(`/trophy/:id`, middleware.isLogged, middleware.isAdmin, middleware.tr
 		
 })
 
+router.post(`/trophyUploadPicture/:id`, middleware.isLogged, middleware.isAdmin, async (req, res) => {
+	
+	console.log("pic trophy upload")
+	const id = req.params.id;
+    //console.log(req.body);
+		
+	const doc = await trophyRef.doc(id).get();	
+	if (!doc.exists) {
+			
+		console.log('No trophy');
+		res.json({errorMessage: `No trophy: ${id}`});
+			
+	} else {
+		
+		console.log(req.body)
+		console.log(req.data)
+		
+		
+		const upload = multer({ storage: multer.memoryStorage() }).single('file')
+		
+		upload(req, res, (err) => {
+		
+			//console.log("Request ---", req.body);
+			console.log("Request file ---", req.file);
+			
+			let bucket = firestore.admin.storage().bucket();
+		
+			try {
+				
+				
+				const blob = bucket.file("trophies/" + doc.id);
+				
+				const blobWriter = blob.createWriteStream({
+					metadata: {
+						contentType: req.file.mimetype
+					}
+				})
+				
+				blobWriter.on('error', (err) => {
+					console.log(err)
+				})
+				
+				blobWriter.on('finish', () => {
+					console.log(`Updated trophy picture with ID: ${id}`);
+					res.status(200).send("File uploaded.")
+				})
+				
+				blobWriter.end(req.file.buffer)
+
+				
+			} catch (error) {
+				console.log("pic error");
+			}
+			
+			
+			//res.json({});
+			
+		})	
+		
+	}
+})
+
 
 // Delete one record
 router.delete(`/trophy/:id`, middleware.isLogged, middleware.isAdmin, async (req, res) => {	
@@ -112,10 +197,90 @@ router.delete(`/trophy/:id`, middleware.isLogged, middleware.isAdmin, async (req
 	const id = req.params.id;
 
 	console.log(`Delete trophy ${req.params.id}`);
-	const del = await trophyRef.doc(id).delete();
+	
+	const docU = await trophyRef.doc(id).get();
+
+    if (!docU.exists) {
+		
+        console.log('No trophy');
+        res.json({errorMessage: `No trophy: ${id}`});
+		
+    }else{
+
+		// img
+		const pic = id;
+		
+		console.log(`Trophy pic: ${pic}`);
+		
+		// delete from storage
+			try {
+				let bucket = firestore.admin.storage().bucket();
+			//console.log(bucket);
+				await bucket.file("trophies/" + pic).delete(); 
+			} catch (error) {
+				console.log("no pic");
+			}
 			
-	res.json({});     
-	   
+		
+			
+		// delete local
+			try { 
+				var fs = require('fs');
+				var filePath = './public/trophies/' + pic;
+				fs.unlink(filePath);
+			} catch (error) {
+				console.log("no local pic");
+			}
+		console.log(`Deleted trophy image`);
+
+		const del = await trophyRef.doc(id).delete();
+        res.json({});    
+    }	   
+})
+
+// remove picture
+router.post(`/trophyDeletePic/:id`, middleware.isLogged, middleware.isAdmin, async (req, res) => {	
+	
+    const id = req.params.id;
+
+
+    const docU = await trophyRef.doc(id).get();
+
+    if (!docU.exists) {
+		
+        console.log('No trophy');
+        res.json({errorMessage: `No trophy: ${id}`});
+		
+    }else{
+
+		// img
+		const pic = id;
+		
+		console.log(`Trophy pic: ${pic}`);
+		
+		// delete from storage
+			try {
+				let bucket = firestore.admin.storage().bucket();
+			//console.log(bucket);
+				await bucket.file("trophies/" + pic).delete(); 
+			} catch (error) {
+				console.log("no pic");
+			}
+			
+		
+			
+		// delete local
+			try { 
+				var fs = require('fs');
+				var filePath = './public/trophies/' + pic;
+				fs.unlink(filePath);
+			} catch (error) {
+				console.log("no local pic");
+			}
+		console.log(`Deleted trophy image`);
+
+        res.json({});    
+    }
 })
 
 module.exports = router
